feat: reorder and improve deployment scripts (#45)
* refactor: renamed DataFeedKeeper to DataFeedStrategy

* feat: added StrategyJob

* test: updated e2e tests

* test: style adjustment

* test: goerli deployment

* refactor: internalized _twapIsOutOfThresholds

* refactor: renaming keeper to strategy

* fix: deployment fixture and test

* fix: dummy force fetch

* feat: reordering deployment scripts and tests

* fix: rm pending tx

* fix: rm .only

* feat: package scripts

* chore: goerli deployments

* fix: chainId type in tests

* feat: fetch bridge and dummy-bridge

* chore: test deployment

* fix: optimisticGoerli name

* fix: tiny fix on dummy deployment

* fix: unit tests

* fix: e2e fixture test after merge

* fix: e2e fixture test

* fix: e2e and unit tests

* fix: yarnlock

* feat: add fixture test expectations

* fix: renaming

* chore: goerli-op deployments

* feat: verify oracle script

* feat: adds dummy same-chain deployment

* chore: mumbai-goerli deployment

* feat: save and verify goerli oracle

* feat: save goerli oracle

* chore: addressing comments in pr

+++ solidity/contracts/DataFeed.sol
@@ -25,7 +25,7 @@ contract DataFeed is IDataFeed, PipelineManagement {
   /// @inheritdoc IDataFeed
   function sendObservations(
     IBridgeSenderAdapter _bridgeSenderAdapter,
-    uint16 _chainId,
+    uint32 _chainId,
     bytes32 _poolSalt,
     uint24 _poolNonce,
     IOracleSidechain.ObservationData[] memory _observationsData
